Add overload, move tests
